Added Get-WAPVMRoleDisk and New-WAPVMRoleVMDisk
Get-WAPVMRoleDisk is similar to the OSDIsk Get without the formatting - essentially ensuring that you can find the data disks ready to attach. I assume most tenancies will have different name schemas so I figured it was easiest just to return them all. 

New-WAPVMRoleVMDisk takes the output of Get and a Cloud Service name, then attachs the disk. Again, the VM has to be stopped, so probably need to put in some form of check there.